feat: added shared files for center and slide modal
